+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 1.0.0.1
 -------
 * removed the upper bounds of the dependencies
+
+1.1.0.0
+-------
+* new function `integratePolynomialOnSimplex`, returning the exact value of 
+  a polynomial over a simplex